LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-eventsource-redirect-to-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to 'http://localhost:8000/eventsource/resources/simple-event-stream.asis' because it violates the following Content Security Policy directive: "connect-src http://127.0.0.1:8000/security/contentSecurityPolicy/resources/redir.php".

PASS EventSource() did not follow the disallowed redirect.
PASS successfullyParsed is true

TEST COMPLETE
